can subscribe to multiple podcasts and also images show up by changing the relationship type in the CoreData model
