+++ cjpodcast/Views/Podcasts/PodcastListItemView.swift
@@ -13,7 +13,7 @@ struct PodcastListItemView: View {
     @Environment(\.managedObjectContext) var managedObjectContext
     @EnvironmentObject var state: PodcastState
     @State var podcast: Podcast
-    
+
     var body: some View {
         HStack {
             Image(uiImage: podcast.image)
@@ -25,51 +25,9 @@ struct PodcastListItemView: View {
                 Text(podcast.publisher).font(.caption).foregroundColor(.gray)
             }
             Spacer()
-            subscribeButton
+            SubscribeButton(podcast: podcast, subscribed: podcast.subscribed)
         }
     }
-    
-    var subscribeButton: some View {
-        Button(action: {
-            print(self.podcast.subscribed)
-            self.podcast.subscribed.toggle()
-            print(self.podcast.subscribed)
-            if self.podcast.subscribed {
-                let newSub = PersistentPodcast(context: self.managedObjectContext)
-                newSub.fromPodcast(podcast: self.podcast)
-
-                do {
-                    try self.managedObjectContext.save()
-                    print("saved")
-                } catch {
-                    print(error)
-                }
-            } else {
-                print("not subbed")
-                do {
-                    let unsubReq = PersistentPodcast.getByTitle(title: self.podcast.title)
-                    let unsubPodcast = try self.managedObjectContext.fetch(unsubReq)
-                    
-                    for pod in unsubPodcast {
-                        self.managedObjectContext.delete(pod)
-                    }
-                    
-                    try self.managedObjectContext.save()
-                } catch {
-                    print(error)
-                }
-            }
-        })
-        {
-            if self.podcast.subscribed {
-                Text("unsubscribe").foregroundColor(.red)
-            } else {
-                Text("subscribe").foregroundColor(.green)
-            }
-        }
-        .buttonStyle(BorderlessButtonStyle())
-    }
-
 }
 
 struct PodcastListItemView_Previews: PreviewProvider {
